Extract MetricsSection with independent Suspense boundary
Give the metrics selector and variant performance chart their own streaming
promise and Suspense boundary with dedicated skeleton and error states.
